libcob:
* common.c (cob_stack_trace_internal): early exit if no data available
* common.c (cob_check_version): test for minimal version 2.2

cobc:
* pplex.l (ppinput): fixed processing after "line not terminated"

[feature-requests:#53] TURN directive and -fno-ec=NAME/-fec=NAME - finish
* cobc.c (turn_ec_for_table): correct handling for EC-ALL
* cobc.c (cobc_turn_ec): handling EC-USER as PENDING
* cobc.c (process_command_line), cobc.h: handle -debug before all the flags and exception options it sets, allowing to remove its internal explicit_set flag

tests:
* run_misc.at: test for [r3957]
* general: pass AWK, GREP, SED to the testsuite and use those
* atlocal.in, atlocal_valgrind, atlocal_win: remove temporary file info.out after use
........
cleanup for 3.1

build_aux:
* create_win_dist.sh: removed obsolete defaults.h
* pre-inst-env.in: exec_prefix as exec_prefix

tests:
* atlocal.in, atlocal_valgrind: exec_prefix as exec_prefix
* testsuite.src/run_misc.at: added test for r3952
........
Adjusted NEWS for 3.1 release

additional: added expected test failure for INITIALIZE OCCURS UNBOUNDED and fixed eol-style attribute for exception-io.def
........
rework file mapping via environment names and apply them to `CBL_`/`C$` file routines

libcob:
* fileio.c (has_directory_separator, looks_absolute): new helpers
* fileio.c (cob_chk_file_mapping, cob_chk_file_env): rewritten:
  * file names / parts that start with period/digit are not converted
  * periods within the name converted to underscores to get the environment variable name
  * empty environment names are not used
  * if COB_FILE_PATH is set, it is always applied if the file name is neither absolute nor start with -F/-D (acu-compat)
  * fileio.c (cob_open): return COB_STATUS_31_INCONSISTENT_FILENAME if it is unset/empty
  * fileio.c (open_cbl_file, cob_sys_delete_file, cob_sys_copy_file, cob_sys_check_file_exist, cob_sys_rename_file): CBL_ and C$ file routines use common mapping rules now
........
handling CHECKNUM directive, preparation for SPZERO

cobc 
* pplex.l, ppparse.y: implemented CHECKNUM/NOCHECKNUM, preparation for SPZERO(NOSPZERO
........
better error-handling for errors in conditional compilation directives

* cobc/ppparse.y: improved error handling for broken IF/ELIF directives by consuming as much errors as possible and emitting a "false"
* tests: moved some compile-only checks for directives from run_extensions.at to syn_misc.at, adding known issues with level 78 constants as expected failures
........
minor cleanup and test adjustments
........
cobc:
* codegen.c (output_initialize_compound): adjusted OCCURS initialization to only do the initialization (and runtime checks) once and then memcpy to all other occurences
* tree.c: preparation for all-upper-case names (post 3.1)

libcob:
* fileio.c: initialize errno in all places where it is checked afterwards
* fileio.c [_WIN32]: implemented file locking via LockFileEx/UnlockFile
* common.c (cob_sys_waitpid) [_WIN32]: fixed logic error in #if, allowing the best process synchronization possible on that platform
* common.h [_MSC_VER]: removed unused global includes

tests:
* atlocal.in, atlocal_valgrind, atlocal_win:
  * fix unsetting of variables which values contain "COB" (rendered the testuite unusable before...)
  * allow screenio-tests to be run on cygwin/msys with ncurses out-of-the box
  * fixed TEST_LOCAL (may not use pre-inst-env) and "external" versions with old indexed file msgid
  * worked around some win32 issues
........

+++ libcob/fsqlxfd.c
@@ -56,7 +56,7 @@ cob_findkey_attr (cob_file *f, cob_field *kf, int *fullkeylen, int *partlen)
 				} else {
 					*partlen = *fullkeylen;
 				}
-				return k;
+				return (int)k;
 			}
 		}
 	}
@@ -104,7 +104,7 @@ db_savekey (cob_file *f, unsigned char *keyarea, unsigned char *record, int idx)
 		return totlen;
 	}
 	memcpy (keyarea, record + f->keys[idx].offset, f->keys[idx].field->size);
-	return f->keys[idx].field->size;
+	return (int)f->keys[idx].field->size;
 }
 
 /* Compare key for given index 'keyarea' to 'record'.
